refactor: move types for revision to separate file
